refactor: replace #ifdef Argus guards with select() split-file pattern
Remove all CAMERA_BACKEND_NVIDIA_ARGUS_ENABLED preprocessor guards
introduced in Phase 2, applying the same select()-without-ifdef strategy
documented in BUILD_CONFIG.md.

Changes:
- encode_pipeline.h/cpp: remove outer #ifdef wrapper (only compiled on
  Argus via select() in BUILD; guard was redundant)
- nvbuf_cuda_utils.h/cpp: same — remove outer #ifdef wrapper
- gpu_frame_processor.h: remove #ifdef guard; drop #include <gst/gst.h>;
  add 'struct GstBuffer;' forward-decl so header compiles on x86 without
  GStreamer; .cpp wrapper also removed (file is Argus-only via select())
- gst_camera_source_impl.cpp: remove GetGpuFrameProcessor() body (moved
  to split files)
  → gst_camera_source_impl_gpu_argus.cpp: real impl with NvidiaArgusBackend
    dynamic_cast; compiled only when --config=nvidia-argus-camera
  → gst_camera_source_impl_gpu_stub.cpp: returns nullptr; compiled otherwise
- bird_watcher.h: add ConnectGpuPath()/DisconnectGpuPath() private decls
- bird_watcher.cpp: replace #ifdef include + Start/Stop blocks with calls
  to ConnectGpuPath()/DisconnectGpuPath(); no more preprocessor in .cpp
  → bird_watcher_gpu_argus.cpp: SetRgbCallback wiring; Argus only
  → bird_watcher_gpu_stub.cpp: no-ops for x86 / V4L2 / stub builds
- adapters/camera/BUILD: remove 'defines' from nvbuf_cuda_utils,
  encode_pipeline, gpu_frame_processor; add select() to gst_camera_source
  srcs choosing _gpu_argus.cpp vs _gpu_stub.cpp
- application/bird_watch/BUILD: remove 'defines'; add select() to srcs
  choosing bird_watcher_gpu_argus.cpp vs bird_watcher_gpu_stub.cpp

Pre-existing #ifdef CAMERA_BACKEND_NVIDIA_ARGUS_ENABLED guards in
gst_camera_source_impl.cpp (constructor) and camera_detector_impl.cpp
are unchanged — those were present before Phase 2.

x86 build: 119/121 pass (same 2 pre-existing failures — GStreamer headers).
