+++ src/cpp_accelerator/application/bird_watch/bird_watcher_gpu_stub.cpp
@@ -1,0 +1,11 @@
+// GPU-path stub for BirdWatcher on non-Argus builds.
+// Compiled on x86 / V4L2 / stub paths (select() in BUILD).
+
+#include "src/cpp_accelerator/application/bird_watch/bird_watcher.h"
+
+namespace jrb::application::bird_watch {
+
+void BirdWatcher::ConnectGpuPath() {}
+void BirdWatcher::DisconnectGpuPath() {}
+
+}  // namespace jrb::application::bird_watch
